requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

demoji==1.1.0
keras==2.11.0
matplotlib==3.6.2
nltk==3.8.1
numpy==1.23.5
pandas==1.5.2
PyWavelets==1.4.1
scikit-learn==1.2.0
# Editable install with no version control (scipy==1.9.3)
seaborn==0.12.2
tensorflow-macos==2.11.0
textblob==0.17.1
transformers==4.25.1
wordcloud==1.8.2.2
yfinance==0.1.90
